Refactor price fields to float and add event/message handling
Replaces BigDecimal price fields with float across Cart, Events, Orders, Payments, and FileGeneration modules for consistency. Adds Feign clients, event/message DTOs, RabbitMQ configuration, and listeners to support inter-service communication and event-driven workflows. Improves validation logic in service layers and error handling for missing resources.

+++ Cart/src/main/java/org/example/dtos/cartItems/CartItemDTO.java
@@ -1,6 +1,5 @@
 package org.example.dtos.cartItems;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -11,7 +10,7 @@ public class CartItemDTO {
 
     private int quantity;
 
-    private BigDecimal unitPrice;
+    private float unitPrice;
 
     private Date createdAt;
