+++ Orders/src/main/java/org/evently/orders/dtos/orderLines/OrderLineDTO.java
@@ -3,7 +3,6 @@ package org.evently.orders.dtos.orderLines;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Getter
@@ -16,6 +15,6 @@ public class OrderLineDTO {
 
     private Integer quantity;
 
-    private BigDecimal unitPrice;
+    private float unitPrice;
 
 }
