+++ Payments/src/main/java/org/example/dtos/PaymentDTO.java
@@ -3,7 +3,6 @@ package org.example.dtos;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -14,7 +13,7 @@ public class PaymentDTO {
     private UUID id;
     private UUID orderId;
     private UUID userId;
-    private BigDecimal amount;
+    private float amount;
     private String status;
     private String provider;
     private String providerRef;
